Custom Exception & Custom Exception Handler
Created custom exception to be thrown from ticket service and to be handled to send back a custom error response

+++ src/main/java/com/YusufFakhreddin/ICDTicketingSystem/service/TicketServiceImpl.java
@@ -1,7 +1,9 @@
 package com.YusufFakhreddin.ICDTicketingSystem.service;
 
+import com.YusufFakhreddin.ICDTicketingSystem.ErrorHandling.CustomException;
 import com.YusufFakhreddin.ICDTicketingSystem.dao.TicketRepo;
 import com.YusufFakhreddin.ICDTicketingSystem.entity.Ticket;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,7 +23,13 @@ public class TicketServiceImpl implements TicketService{
 
     @Override
     public Ticket getTicket(String id){
-        return ticketRepo.findById(id).get();
+        // check for id existence and throw custom error if not
+        Ticket ticket = ticketRepo.findById(id).orElse(null);
+        if (ticket == null) {
+            throw new CustomException(HttpStatus.NOT_FOUND.value(),"Ticket id not found - " + id);
+        }
+        return ticket;
+
     }
 
     @Override
@@ -30,10 +38,24 @@ public class TicketServiceImpl implements TicketService{
     }
 
     @Override
-    public Ticket updateTicket(Ticket ticket){return ticketRepo.save(ticket);}
+    public Ticket updateTicket(String id , Ticket ticket){
+        // check for id existence and throw custom error if not
+        if (ticketRepo.findById(id).orElse(null) == null) {
+            throw new CustomException(HttpStatus.NOT_FOUND.value(),"Ticket id not found - " + id);
+        }
+        // append id to ticket object
+        ticket.setId(id);
+        return ticketRepo.save(ticket);
+    }
 
     @Override
     public void deleteTicket(String id){
+
+// check for id existence and throw custom error if not
+        if (ticketRepo.findById(id).orElse(null) == null) {
+            throw new CustomException(HttpStatus.NOT_FOUND.value(),"Ticket id not found - " + id);
+        }
+
         ticketRepo.deleteById(id);
     }
 }
